fix: 3C 피드 페이지 페이지  구성 및 레이아웃 수정 3C 피드 페이지 컨텐츠 개별 컴포넌트로 분리(ThumbnailFeedList) -사용자 프로필의 코드 재사용을 위함 열마다 있던 여백 제거 페이지 반응형 레이아웃 변경- lg:4 -> lg:8 (메인 피드 페이지의 2배) Fixes: #14
