+++ src/pages/feed/css/ThumbnailFeedList.module.css
@@ -1,0 +1,24 @@
+.thumbContainer {
+    display: flex;
+    flex-wrap: wrap;
+    align-content: flex-start;
+    justify-content: center;
+    align-items: center;
+    width: 100%;
+    height: 100vh;
+    margin: 0 auto;
+}
+
+.thumbRow {
+    width: 100%;
+    height: fit-content;
+}
+
+.emptyCol {
+    width: 100%;
+    position: relative;
+    padding-top: 100%;
+    overflow: hidden;
+    /* background-color: gray; */
+    border-radius: 5px;
+}
